requirements.txt
fastapi
uvicorn
pillow
numpy 